tasky/ocr_ia/requirements.txt
fastapi
uvicorn
pillow
requests
pytesseract
psycopg2-binary
